Implemented showing label with frequency-amplitude data on mouse hover on FFT view;

+++ VoiceChanger/SpectrumCreator/FrequencyAmplitudeData.cs
@@ -7,13 +7,13 @@
 
         }
 
-        public FrequencyAmplitudeData(int frequency, float amplitude)
+        public FrequencyAmplitudeData(float frequency, float amplitude)
         {
             Frequency = frequency;
             Amplitude = amplitude;
         }
 
-        public int Frequency { get; set; }
+        public float Frequency { get; set; }
         public float Amplitude { get; set; }
 
         public override string ToString()
